requirements.txt
numpy
tensorflow
keras
Pillow
scikit-learn
matplotlib
piexif
